Move SystemInfo proto to protos/perfetto/common
