+++ include/perfetto/ext/base/android_utils.h
@@ -22,12 +22,6 @@
 
 #include "perfetto/base/build_config.h"
 
-#if PERFETTO_BUILDFLAG(PERFETTO_OS_ANDROID) || \
-    PERFETTO_BUILDFLAG(PERFETTO_OS_LINUX) ||   \
-    PERFETTO_BUILDFLAG(PERFETTO_OS_APPLE)
-#include <sys/utsname.h>
-#endif
-
 namespace perfetto {
 namespace base {
