+++ protos/perfetto/common/system_info.proto
