+++ protos/perfetto/ipc/BUILD.gn
@@ -33,7 +33,6 @@ perfetto_proto_library("@TYPE@") {
     # .ipc.h depends on the .gen.h coming from the various common/xxx.proto.
     "../common:cpp",
     "../config:cpp",
-    "../trace:minimal_cpp",
   ]
 }
