+++ protos/perfetto/trace/BUILD.gn
@@ -42,7 +42,6 @@ proto_sources_minimal = [
   "clock_snapshot.proto",
   "trace_uuid.proto",
   "trigger.proto",
-  "system_info.proto",
 ]
 
 # Most targets should either depend on :zero (writers) / :lite (readers)
