+++ protos/perfetto/trace/perfetto_trace.proto
@@ -4676,6 +4676,73 @@ message TraceConfig {
 
 // End of protos/perfetto/config/trace_config.proto
 
+// Begin of protos/perfetto/common/system_info.proto
+
+message Utsname {
+  optional string sysname = 1;
+  optional string version = 2;
+  optional string release = 3;
+  optional string machine = 4;
+}
+
+// Next id: 15;
+message SystemInfo {
+  optional Utsname utsname = 1;
+  optional string android_build_fingerprint = 2;
+
+  // The manufacturer of the product/hardware.
+  // Source : "ro.product.manufacturer"
+  // Introduced after Android W in Nov 2024 and is not supported on older
+  // versions.
+  optional string android_device_manufacturer = 14;
+
+  // The SoC model from which trace is collected
+  optional string android_soc_model = 9;
+
+  // The guest SoC model from which trace is collected in case of VMs
+  optional string android_guest_soc_model = 13;
+
+  // The hardware reversion from android device
+  optional string android_hardware_revision = 10;
+
+  // The storage component from android_device. This field has been introduced
+  // after Android W in Aug 2024 and is not supported on older versions.
+  optional string android_storage_model = 11;
+
+  // The RAM component information from android device. This field has been
+  // introduced after Android W in Aug 2024 and is not supported on older
+  // versions.
+  optional string android_ram_model = 12;
+
+  // The version of traced (the same returned by `traced --version`).
+  // This is a human readable string with and its format varies depending on
+  // the build system and the repo (standalone vs AOSP).
+  // This is intended for human debugging only.
+  optional string tracing_service_version = 4;
+
+  // The Android SDK vesion (e.g. 21 for L, 31 for S etc).
+  // Introduced in Android T.
+  optional uint64 android_sdk_version = 5;
+
+  // Kernel page size - sysconf(_SC_PAGESIZE).
+  optional uint32 page_size = 6;
+
+  // Number of cpus - sysconf(_SC_NPROCESSORS_CONF).
+  // Might be different to the number of online cpus.
+  // Introduced in perfetto v44.
+  optional uint32 num_cpus = 8;
+
+  // The timezone offset from UTC, as per strftime("%z"), in minutes.
+  // Introduced in v38 / Android V.
+  optional int32 timezone_off_mins = 7;
+
+  // Ticks per second - sysconf(_SC_CLK_TCK).
+  // Not serialised as of perfetto v44.
+  optional int64 hz = 3;
+}
+
+// End of protos/perfetto/common/system_info.proto
+
 // Begin of protos/perfetto/common/trace_stats.proto
 
 // Statistics for the internals of the tracing service.
@@ -15254,73 +15321,6 @@ message SysStats {
 }
 
 // End of protos/perfetto/trace/sys_stats/sys_stats.proto
-
-// Begin of protos/perfetto/trace/system_info.proto
-
-message Utsname {
-  optional string sysname = 1;
-  optional string version = 2;
-  optional string release = 3;
-  optional string machine = 4;
-}
-
-// Next id: 15;
-message SystemInfo {
-  optional Utsname utsname = 1;
-  optional string android_build_fingerprint = 2;
-
-  // The manufacturer of the product/hardware.
-  // Source : "ro.product.manufacturer"
-  // Introduced after Android W in Nov 2024 and is not supported on older
-  // versions.
-  optional string android_device_manufacturer = 14;
-
-  // The SoC model from which trace is collected
-  optional string android_soc_model = 9;
-
-  // The guest SoC model from which trace is collected in case of VMs
-  optional string android_guest_soc_model = 13;
-
-  // The hardware reversion from android device
-  optional string android_hardware_revision = 10;
-
-  // The storage component from android_device. This field has been introduced
-  // after Android W in Aug 2024 and is not supported on older versions.
-  optional string android_storage_model = 11;
-
-  // The RAM component information from android device. This field has been
-  // introduced after Android W in Aug 2024 and is not supported on older
-  // versions.
-  optional string android_ram_model = 12;
-
-  // The version of traced (the same returned by `traced --version`).
-  // This is a human readable string with and its format varies depending on
-  // the build system and the repo (standalone vs AOSP).
-  // This is intended for human debugging only.
-  optional string tracing_service_version = 4;
-
-  // The Android SDK vesion (e.g. 21 for L, 31 for S etc).
-  // Introduced in Android T.
-  optional uint64 android_sdk_version = 5;
-
-  // Kernel page size - sysconf(_SC_PAGESIZE).
-  optional uint32 page_size = 6;
-
-  // Number of cpus - sysconf(_SC_NPROCESSORS_CONF).
-  // Might be different to the number of online cpus.
-  // Introduced in perfetto v44.
-  optional uint32 num_cpus = 8;
-
-  // The timezone offset from UTC, as per strftime("%z"), in minutes.
-  // Introduced in v38 / Android V.
-  optional int32 timezone_off_mins = 7;
-
-  // Ticks per second - sysconf(_SC_CLK_TCK).
-  // Not serialised as of perfetto v44.
-  optional int64 hz = 3;
-}
-
-// End of protos/perfetto/trace/system_info.proto
 
 // Begin of protos/perfetto/trace/system_info/cpu_info.proto
