+++ src/base/android_utils.cc
@@ -32,6 +32,7 @@
 #if PERFETTO_BUILDFLAG(PERFETTO_OS_ANDROID) || \
     PERFETTO_BUILDFLAG(PERFETTO_OS_LINUX) ||   \
     PERFETTO_BUILDFLAG(PERFETTO_OS_APPLE)
+#include <sys/utsname.h>
 #include <unistd.h>
 #endif
